LLM/Deploy/requirements.txt
llama-cpp-python
langchain
fastapi
uvicorn